+++ Include/Acidanthera/Library/OcConfigurationLib.h
@@ -344,7 +344,7 @@ OC_DECLARE (OC_MISC_BLESS_ARRAY)
   _(BOOLEAN                     , PickerAudioAssist           ,     , FALSE                               , ())                   \
   _(BOOLEAN                     , HideAuxiliary               ,     , FALSE                               , ())                   \
   _(BOOLEAN                     , PollAppleHotKeys            ,     , FALSE                               , ())                   \
-  _(BOOLEAN                     , ShowPicker                  ,     , FALSE                               , ())
+  _(OC_STRING                   , ShowPicker                  ,     , OC_STRING_CONSTR ("Always", _, __)  , OC_DESTR (OC_STRING) )
 OC_DECLARE (OC_MISC_BOOT)
 
 #define OC_MISC_DEBUG_FIELDS(_, __) \